[Add] CRC value into data record

+++ internal/sstable/HelperStructures.go
@@ -3,7 +3,7 @@ package sstable
 import "time"
 
 type DataElement struct {
-	CRC       uint64
+	CRC       uint32
 	Timestamp time.Time
 	Tombstone bool
 	KeySize   uint64
